groupmaker
Updated groupmaker, there were two issues. #1: The ampersand was messing up the readings. So it now filtersout ampersands in titles. It had been using the &amp; from html. #2: It was basing its assignments from the new synergy exports based on the first row of students. So if there was a blank it would stop. I changed that, now it does it from the 3rd column of assigments (which is 3)